move loader class to same file as v1 loader
