+++ src/nplinker/genomics/bigscape/__init__.py
@@ -1,6 +1,6 @@
 import logging
 from .bigscape_loader import BigscapeGCFLoader
-from .bigscape_loader_v2 import BigscapeV2GCFLoader
+from .bigscape_loader import BigscapeV2GCFLoader
 from .runbigscape import run_bigscape
